refactor(json-schema-2020-12): consolidate existing code + styles
Refs #8513
