+++ src/core/plugins/json-schema-2020-12/components/JSONSchema/_json-schema.scss
@@ -32,31 +32,12 @@
     border-radius: 4px;
   }
 
-  &-core-keyword {
-    color: #6b6b6b;
-    font-size: 12px;
-    font-style: italic;
-    margin-left: 20px;
-    font-weight: bold;
-
-    &__value {
-      color: #6b6b6b;
-      font-style: italic;
-      font-size: 12px;
-      font-weight: normal;
-    }
-  }
-
   //&-note {
   //  @include text_headline($section-models-model-title-font-color);
   //  padding: 10px 0 0 20px;
   //  font-size: 11px;
   //  color: #6b6b6b;
   //}
-}
-
-.json-schema-2020-12-core-keyword + .json-schema-2020-12-core-keyword__value::before {
-  content: '='
 }
