+++ src/core/plugins/json-schema-2020-12/components/_all.scss
@@ -6,18 +6,4 @@
 @import './JSONSchema/json-schema';
 @import './Accordion/accordion';
 @import './ExpandDeepButton/expand-deep-button';
-@import './keywords/$vocabulary/$vocabulary';
-@import './keywords/$defs/$defs';
-@import './keywords/AllOf/all-of';
-@import './keywords/AnyOf/any-of';
-@import './keywords/OneOf/one-of';
-@import './keywords/Not/not';
-@import './keywords/If/if';
-@import './keywords/Then/then';
-@import './keywords/Else/else';
-@import './keywords/DependentSchemas/dependent-schemas';
-@import './keywords/Type/type';
-@import './keywords/Format/format';
-@import './keywords/Description/description';
-@import './keywords/Title/title';
-@import './keywords/Properties/properties';
+@import './keywords/all';
