--- src/core/plugins/json-schema-2020-12/components/keywords/$defs/_$defs.scss
@@ -1,12 +1,0 @@
-.json-schema-2020-12 {
-  &__\$defs {
-    & ul {
-      @include expansion-border;
-      padding: 0;
-    }
-  }
-
-  &-\$def {
-    @extend .json-schema-2020-12-property;
-  }
-}
