+++ src/core/plugins/json-schema-2020-12/components/keywords/$dynamicRef.jsx
@@ -9,9 +9,11 @@ const $dynamicRef = ({ schema }) => {
   if (!schema?.$dynamicRef) return null
 
   return (
-    <div className="json-schema-2020-12__$dynamicRef">
-      <span className="json-schema-2020-12-core-keyword">$dynamicRef</span>
-      <span className="json-schema-2020-12-core-keyword__value">
+    <div className="json-schema-2020-12-keyword json-schema-2020-12-keyword--$dynamicRef">
+      <span className="json-schema-2020-12-keyword__name json-schema-2020-12-keyword__name--secondary">
+        $dynamicRef
+      </span>
+      <span className="json-schema-2020-12-keyword__value json-schema-2020-12-keyword__value--secondary">
         {schema.$dynamicRef}
       </span>
     </div>
