+++ src/core/plugins/json-schema-2020-12/components/keywords/$id.jsx
@@ -9,9 +9,11 @@ const $id = ({ schema }) => {
   if (!schema?.$id) return null
 
   return (
-    <div className="json-schema-2020-12__$id">
-      <span className="json-schema-2020-12-core-keyword">$id</span>
-      <span className="json-schema-2020-12-core-keyword__value">
+    <div className="json-schema-2020-12-keyword json-schema-2020-12-keyword--$id">
+      <span className="json-schema-2020-12-keyword__name json-schema-2020-12-keyword__name--secondary">
+        $id
+      </span>
+      <span className="json-schema-2020-12-keyword__value json-schema-2020-12-keyword__value--secondary">
         {schema.$id}
       </span>
     </div>
