+++ src/core/plugins/json-schema-2020-12/components/keywords/$schema.jsx
@@ -9,9 +9,11 @@ const $schema = ({ schema }) => {
   if (!schema?.$schema) return null
 
   return (
-    <div className="json-schema-2020-12__$schema">
-      <span className="json-schema-2020-12-core-keyword">$schema</span>
-      <span className="json-schema-2020-12-core-keyword__value">
+    <div className="json-schema-2020-12-keyword json-schema-2020-12-keyword--$schema">
+      <span className="json-schema-2020-12-keyword__name json-schema-2020-12-keyword__name--secondary">
+        $schema
+      </span>
+      <span className="json-schema-2020-12-keyword__value json-schema-2020-12-keyword__value--secondary">
         {schema.$schema}
       </span>
     </div>
