--- src/core/plugins/json-schema-2020-12/components/keywords/AllOf/_all-of.scss
@@ -1,19 +1,0 @@
-.json-schema-2020-12 {
-  &__allOf {
-    margin: 5px 0 5px 0;
-
-    & > ul {
-      @include expansion-border;
-      padding: 0;
-    }
-  }
-
-  &-core-keyword {
-    &--allOf {
-      color: $text-code-default-font-color;
-      font-style: normal;
-    }
-  }
-}
-
-
