--- src/core/plugins/json-schema-2020-12/components/keywords/AnyOf/_any-of.scss
@@ -1,13 +1,0 @@
-.json-schema-2020-12 {
-  &__anyOf {
-    @extend .json-schema-2020-12__allOf;
-  }
-
-  &-core-keyword {
-    &--anyOf {
-      @extend .json-schema-2020-12-core-keyword--allOf;
-    }
-  }
-}
-
-
