--- src/core/plugins/json-schema-2020-12/components/keywords/DependentSchemas/_dependent-schemas.scss
@@ -1,13 +1,0 @@
-.json-schema-2020-12 {
-  &__dependentSchemas {
-    @extend .json-schema-2020-12__allOf;
-  }
-
-  &-core-keyword {
-    &--dependentSchemas {
-      @extend .json-schema-2020-12-core-keyword--allOf;
-    }
-  }
-}
-
-
