+++ src/core/plugins/json-schema-2020-12/components/keywords/Description/Description.jsx
@@ -9,7 +9,11 @@ const Description = ({ schema }) => {
   if (!schema?.description) return null
 
   return (
-    <div className="json-schema-2020-12__description">{schema.description}</div>
+    <div className="json-schema-2020-12-keyword json-schema-2020-12-keyword--description">
+      <div className="json-schema-2020-12-core-keyword__value json-schema-2020-12-core-keyword__value--secondary">
+        {schema.description}
+      </div>
+    </div>
   )
 }
