+++ src/core/plugins/json-schema-2020-12/components/keywords/Description/_description.scss
@@ -1,4 +1,4 @@
-.json-schema-2020-12__description {
+.json-schema-2020-12-keyword--description {
   color: #6b6b6b;
   font-size: 12px;
   margin-left: 20px;
