--- src/core/plugins/json-schema-2020-12/components/keywords/Else/_else.scss
@@ -1,10 +1,0 @@
-.json-schema-2020-12 {
-  &__else {
-    .json-schema-2020-12-core-keyword--else {
-      @extend .json-schema-2020-12-core-keyword--allOf;
-    }
-  }
-}
-
-
-
