+++ src/core/plugins/json-schema-2020-12/components/keywords/If.jsx
@@ -3,21 +3,21 @@
  */
 import React from "react"
 
-import { schema } from "../../../prop-types"
-import { useComponent } from "../../../hooks"
+import { schema } from "../../prop-types"
+import { useComponent } from "../../hooks"
 
 const If = ({ schema }) => {
   if (!schema?.if) return null
 
   const JSONSchema = useComponent("JSONSchema")
   const name = (
-    <span className="json-schema-2020-12-core-keyword json-schema-2020-12-core-keyword--if">
+    <span className="json-schema-2020-12-keyword__name json-schema-2020-12-keyword__name--primary">
       If
     </span>
   )
 
   return (
-    <div className="json-schema-2020-12__if">
+    <div className="json-schema-2020-12-keyword json-schema-2020-12-keyword--if">
       <JSONSchema name={name} schema={schema.if} />
     </div>
   )
