--- src/core/plugins/json-schema-2020-12/components/keywords/If/_if.scss
@@ -1,10 +1,0 @@
-.json-schema-2020-12 {
-  &__if {
-    .json-schema-2020-12-core-keyword--if {
-      @extend .json-schema-2020-12-core-keyword--allOf;
-    }
-  }
-}
-
-
-
