--- src/core/plugins/json-schema-2020-12/components/keywords/Not/_not.scss
@@ -1,10 +1,0 @@
-.json-schema-2020-12 {
-  &__not {
-    .json-schema-2020-12-core-keyword--not {
-      @extend .json-schema-2020-12-core-keyword--allOf;
-    }
-  }
-}
-
-
-
