--- src/core/plugins/json-schema-2020-12/components/keywords/OneOf/_one-of.scss
@@ -1,13 +1,0 @@
-.json-schema-2020-12 {
-  &__oneOf {
-    @extend .json-schema-2020-12__allOf;
-  }
-
-  &-core-keyword {
-    &--oneOf {
-      @extend .json-schema-2020-12-core-keyword--allOf;
-    }
-  }
-}
-
-
