--- src/core/plugins/json-schema-2020-12/components/keywords/Then/_then.scss
@@ -1,10 +1,0 @@
-.json-schema-2020-12 {
-  &__then {
-    .json-schema-2020-12-core-keyword--then {
-      @extend .json-schema-2020-12-core-keyword--allOf;
-    }
-  }
-}
-
-
-
