+++ src/core/plugins/json-schema-2020-12/components/keywords/Title/_title.scss
@@ -4,12 +4,10 @@
     display: inline-block;
     font-weight: bold;
 
-    & .json-schema-2020-12-core-keyword {
+    & .json-schema-2020-12-keyword__name {
       margin: 0;
     }
   }
-
-
 
   &-property {
     margin: 7px 0;
