+++ src/core/plugins/json-schema-2020-12/hoc.jsx
@@ -11,16 +11,16 @@ import Keyword$anchor from "./components/keywords/$anchor"
 import Keyword$dynamicAnchor from "./components/keywords/$dynamicAnchor"
 import Keyword$ref from "./components/keywords/$ref"
 import Keyword$dynamicRef from "./components/keywords/$dynamicRef"
-import Keyword$defs from "./components/keywords/$defs/$defs"
+import Keyword$defs from "./components/keywords/$defs"
 import Keyword$comment from "./components/keywords/$comment"
-import KeywordAllOf from "./components/keywords/AllOf/AllOf"
-import KeywordAnyOf from "./components/keywords/AnyOf/AnyOf"
-import KeywordOneOf from "./components/keywords/OneOf/OneOf"
-import KeywordNot from "./components/keywords/Not/Not"
-import KeywordIf from "./components/keywords/If/If"
-import KeywordThen from "./components/keywords/Then/Then"
-import KeywordElse from "./components/keywords/Else/Else"
-import KeywordDependentSchemas from "./components/keywords/DependentSchemas/DependentSchemas"
+import KeywordAllOf from "./components/keywords/AllOf"
+import KeywordAnyOf from "./components/keywords/AnyOf"
+import KeywordOneOf from "./components/keywords/OneOf"
+import KeywordNot from "./components/keywords/Not"
+import KeywordIf from "./components/keywords/If"
+import KeywordThen from "./components/keywords/Then"
+import KeywordElse from "./components/keywords/Else"
+import KeywordDependentSchemas from "./components/keywords/DependentSchemas"
 import KeywordProperties from "./components/keywords/Properties/Properties"
 import KeywordType from "./components/keywords/Type/Type"
 import KeywordFormat from "./components/keywords/Format/Format"
